replace fixtures in conftest
